relocate visualization code
